Humble developing

+++ obj_to_pointcloud/src/obj_to_pointcloud.cpp
@@ -27,23 +27,22 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <ros/ros.h>
-#include <sensor_msgs/PointCloud2.h>
-#include <pcl_ros/point_cloud.h>
-#include <pcl_ros/transforms.h>
-
-#include <pcl/point_types.h>
-#include <pcl/conversions.h>
-#include <pcl_conversions/pcl_conversions.h>
-#include <pcl/io/vtk_lib_io.h>
-#include <pcl/filters/voxel_grid.h>
-
 #include <cmath>
-#include <random>
-#include <string>
 #include <iostream>
+#include <random>
 #include <sstream>
+#include <string>
 #include <vector>
+
+#include <ros/ros.h>
+
+#include <sensor_msgs/PointCloud2.h>
+
+#include <pcl/filters/voxel_grid.h>
+#include <pcl/io/vtk_lib_io.h>
+#include <pcl/point_cloud.h>
+#include <pcl/point_types.h>
+#include <pcl_conversions/pcl_conversions.h>
 
 #include <neonavigation_common/compatibility.h>
 
@@ -199,10 +198,11 @@ private:
           auto a = p1 - p0;
           auto b = p2 - p0;
 
-          float s = 0.5 * sqrtf(
-                              powf(a.y * b.z - a.z * b.y, 2.0) +
-                              powf(a.z * b.x - a.x * b.z, 2.0) +
-                              powf(a.x * b.y - a.y * b.x, 2.0));
+          float s =
+              0.5 * std::sqrt(
+                        std::pow(a.y * b.z - a.z * b.y, 2) +
+                        std::pow(a.z * b.x - a.x * b.z, 2) +
+                        std::pow(a.x * b.y - a.y * b.x, 2));
           float numf = ppmsq_ * s;
           int num = numf;
           if (numf - num > ud(engine_))
